Search: Improves search resilience
Elasticsearch is not guaranteed to be in sync with our database. The
server might also not be available or all data may be temporarily gone.

We have of course been aware of this and have counter-measures in the
code. However there were some loopholes which should now be closed.

Apart from the search not being available, the user should not see
any error messages anymore at this point.

TYPE: Other