Core: No longer delete cookies
This fixes an issue with logout messages never being shown to the user.
Security wise this does not make a difference, because when we forget a
user we do so by removing the necessary authentication bits in our
server-side session cache. That is, if a certain session id stored in
the cookie is deleted, it is irrelevant if that session id is requested
again, we have already downgraded it.

TYPE: Other