Feriennet: Adds locations for inline-loaded activites
In the activity list, clicking on "more activites" did not use any
browser history yet. This caused issues with Chrome which does not come
with a bfcache and therefore tries to re-load the site, losing all
navigational state.

This change solves this issue by updating the URL for each loaded page.

TYPE: Feature
LINK: FER-756