Adds loading of custom fonts
New fonts can be included by placing them as scss files in `theme/fonts`.
The font family for a file is derived from it's filename which must be congruent
to the css syntax.
Fonts are accessible via the general-settings view.
They are not lazy loaded, but will be cached by the browser.
Theme is compiled if options are passed to the compile function.