Form: Extends pricing options with a credit card payment flag.
This allows users to define custom forms where the credit card payment
is mandatory based on whether or not the person submitting the form has
selected a certain option or not.

E.g. a delivery field where they have to pay online right away if they
want the item to be delivered, but they can pay later in person if
they choose to pick it up themselves instead:

```
Delivery *=
    ( ) Pickup (0 CHF)
    ( ) Delivery (5 CHF!)
```

TYPE: Feature
LINK: OGC-910