Mypy: Adds static type checking to CI/pre-commit (#828)
mypy: Enables static type checking

This is a very lax base config and does not add any annotations beyond
the bare minimum to make mypy happy. But this allows us to get started
on adding type annotations to our code and gradually make the config
more strict for the parts of the code that is annotated.