Adds a slightly faster CI workflow that will be used most of the time
Rather than running tests on the freshly built container and introducing
a dependency between steps this runs tests on `onegov_cloud:head` and
replaces the install from `/app` with the one from the current commit on
`/src`, so the tests in most cases will still show an accurate picture.

Release builds will still use the older, slower workflow and you can
force the CI to use the old workflow if you include [test-container]
in your commit message. In case you need to debug failures that are
specific to the new version of the container.