Fix: use fresh session, explicitely commit transactions.
In some cases, websocket-server was not able to find the Chats table
because the schema was not present in this session. This seemed to be
the case only for cached sessions (using `SESSIONS`). Using a fresh
session obtained from session_manager works, because session_manager
explicitely configures schema for each session.

Outside of a morepath environment, transactions are not automatically
committed (e.g., after a request). Hence, chat has had some open
transactions in idle state. Explicitely commit()ing closes the
transaction and avoids eventual problems with postgres closing the
connections.