Org: Allows uploaded files to appear in public search results more often
Previously only files marked as a publication would appear in search
results, but now it will also check for any public pages linking to the
file in which case it will appear in the search results as well.

This also creates an explicit link for any files linked within a page's
text.

TYPE: Feature
LINK: OGC-921