Org: Implement logic to prevent premature archiving of reservations.
TYPE: Bugfix
LINK: OGC-1481

Imagine a reservation made a year in advance (which happens in practice)
After a year, `ticket.last_change` would indicate it shall be archived.
However, some reservations of this ticket might be fairly recent, like a
month ago. Therefore it is a bit premature to be considered for archive.