Org: Bugfix of deleting files
Some files in the FileLinkExtension could not be deleted, they will return to the list if they are still linked in the text. Some files were invisibly linked in the text. This change will delete all invisible links on save and adds a command for deleting all current invisible links.

TYPE: Bugfix
LINK: PRO-1248