Use create-react-class to wrap legacy React.createClass API.
create-react-class is a wrapper for the legacy React.createClass API.
Using create-react-class, we can continue to use the deprecated way of
creating React components - otherwise, we have to upgrade each component
manually, a time and error-prone process.

For now, we can use latest React without changing our components.
However, we should update a component whenever we work on it to the
latest format.