Core: Increases security of some core constructs
This also includes a fix for `session_id`'s not being properly rotated
when they become invalid. Generally the security was high enough for
what we were using these constructs for, but this may change in the
future, so it's better to have them be as robust as possible now.

TYPE: Bugfix
LINK: SEA-1051