Search: Avoids Postgres indexer causing invalid transactions
This problem only manifested itself in large import jobs where a lot of
ORM events are being generated and the indexer has to be called in the
middle of a transaction, rather than at the end. Since we don't yet use
the Postgres index we haven't fully fixed this yet and instead drop the
ORM events we can't fit into our queue.

This also fixes `ensure_user` failing if the new `username` is already
taken by another user.

TYPE: Bugfix
LINK: OGC-1400