Fixes release script parsing issue of remote tag.
The release script was failing with the following:
`do/release: line 150: test: unbound variable`

The reason being that some remote tag (pushed on
another branch) contained the release string
"release-2024.53-test-1"

The script tried to increment this release, but
failed, because it contains the string 'test'. To
solve this, we now only consider tags in master,
so that experiments do not affect us here.