Fixes `do/release` being broken with newer versions of `git`
Because `head` exits from `git tag` after the first line is passed
when `git` tries to push additional lines into the pipe it will receive
a `SIGPIPE` for a broken pipe and exit with a nonzero exit code, which
causes the entire script to fail due to `-e` and `-o pipefail`.

The simple solution is to reverse the sort order and use `tail` instead.