Created a test which caputres this issue precisely
This would have failed before the fix. The way we can test this,
is inject ourselves into the sqlalchemy update events of order
and count them.

First I attempted to test it by just counting the amount of
seconds it takes, but this ultimately wasn't the best path, as
it may give inconsistent results (e.g. flaky tests).