Server: Avoids caching partially configured application instances
Since `CachedApplication` was setting `self.instance` before the
application was configured and the server will continue serving
subsequent requests after a failed one, it was possible for a
partially initialized application to end up in the cache, which
leads to cryptic error messages.

It's better if we fail the same way on every request, so we don't
get distracted by red herrings.

TYPE: Bugfix