Org: Hides tickets that can't be processed from tickets list
Previously they were displayed, but couldn't be opened, because the
filter happened after the query, so completely hiding would've broken
pagination.

We now do our best to apply the filter as part of the query, so we get
a correct pagination, total number of results etc.

This also fixes a corner-case with non-exclusive ticket permissions on
a ticket group vs. exclusive ticket permissions on the handler code.

TYPE: Bugfix
LINK: OGC-2459