Translator Directory: Link time report emails to ticket status page.
Previously, time report emails sent to translators contained links that
pointed directly to the admin ticket page, which translators couldn't
access. This changes the links to point to the public ticket status page
instead. This allows them to respond with a message.
The email also contains a reply-to email now.

TYPE: Bugfix